+++ core/src/infrastructure/Sha256.cpp
@@ -27,7 +27,7 @@ std::uint32_t RotateRight(std::uint32_t value, unsigned int bits)
 std::vector<std::uint8_t> Pad(std::span<const std::uint8_t> message)
 {
     std::vector<std::uint8_t> padded(message.begin(), message.end());
-    const std::uint64_t bitLength = static_cast<std::uint64_t>(message.size()) * 8;
+    const std::uint64_t bitLength = message.size() * 8;
 
     padded.push_back(0x80);
     while (padded.size() % 64 != 56)
